refactor(vm): split config.rs into a module per config section
config.rs still carried three `// ===` section dividers after the
dead-code deletion — jailer config, top-level VMM config, and the
snapshot-type vocabulary — the exact signal CLAUDE.md calls out as
requiring a module split, not a monolithic file.

Split into config/jailer.rs, config/vmm.rs, and config/snapshot.rs,
with config.rs reduced to module declarations and the crate's existing
re-export surface (crate::config::JailerConfig etc. and the lib.rs
pub use both unchanged). No behavior change; the same 219 unit tests
move with their section.

Found by review (greptile-apps) on PR #626.
